requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cohere==4.17.0
deprecation==2.1.0
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.2
frozenlist==1.4.0
gotrue==1.0.4
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
loguru==0.7.0
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pinecone-client==2.2.2
postgrest==0.10.8
pyarrow==12.0.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
realtime==1.0.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
websockets==10.4
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
